Project mode, configuration added, split files
Sawsge now operates on the directory supplied as the first argument. The
out directory no longer builds in the working directory, but in the
directory supplied as the argument. A configuration file must now be
present in the source directory (an example is in the repo). There are
now two supported modes, project and blog. Project uses pages (no dates,
no summary) for markdown files, while blog uses a combination of posts
(date and summary) and a homepage. Lastly, classes were split into
separate files for organization.