Refactor entire API
Sawsge can now be better used as a library. This changes the Sawsge
module to a class to be instantiated with a project config of the user's
choice. The project can then be built by calling `my_project.build`.
Alternatively, Sawsge can be used with command line arguments with
`Sawsge.cli`. This automatically creates a config from the first
argument provided or the current working directory and build the
project.

Ergo, project configuration is no longer done via module globals. A new
file `lib/sawsge/config.rb` contains a new class to store configuration
options.

This change replaces the `toml` library with the `tomlrb` library.
Tomlrb has superior performance and allows symbolization of result keys.

`@resource_paths` was changed from a Set to an Array to make refactoring
easier. This may be reverted. Additionally, I've removed the old way of
creating the basic body of a document using the `add_next_sibling`
function of `Nokogiri::HTML5`. Now, the header, body, and footer are
simply concatenated. It may not be the 'right' way to do things, but it
is simpler and more readable.